App/requirements.txt
alembic==1.4.0
blinker==1.4
Bootstrap-Flask==1.2.0
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
MarkupSafe==1.1.1
pymongo==3.10.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
python-dotenv==0.14.0
